2/2 Much needed refactor
