+++ reasoner/v4/ActorNode.java
@@ -1,5 +1,6 @@
 package com.vaticle.typedb.core.reasoner.v4;
 
+import com.vaticle.typedb.core.common.exception.TypeDBException;
 import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.common.iterator.Iterators;
 import com.vaticle.typedb.core.concurrent.actor.Actor;
@@ -9,6 +10,8 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
+
+import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 
 public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE> {
 
@@ -44,7 +47,43 @@ public abstract class ActorNode<NODE extends ActorNode<NODE>> extends Actor<NODE
     // TODO: Since port has the index in it, maybe we don't need index here?
     public abstract void readAnswerAt(ActorNode.Port reader, int index);
 
-    public abstract void receive(ActorNode.Port port, Message message);
+
+
+    public void receive(Port onPort, Message received) {
+        switch (received.type()) {
+            case ANSWER: {
+                handleAnswer(onPort, received.asAnswer());
+                break;
+            }
+            case CONCLUSION: {
+                handleConclusion(onPort, received.asConclusion());
+                break;
+            }
+            case CONDITIONALLY_DONE: {
+                handleConditionallyDone(onPort);
+                break;
+            }
+            case DONE: {
+                handleDone(onPort);
+                break;
+            }
+            default:
+                throw TypeDBException.of(ILLEGAL_STATE);
+        }
+    }
+
+    protected abstract void handleAnswer(Port onPort, Message.Answer answer);
+
+    protected void handleConclusion(Port onPort, Message.Conclusion conclusion) {
+        throw TypeDBException.of(ILLEGAL_STATE);
+    }
+
+    protected void handleConditionallyDone(Port onPort) {
+        throw TypeDBException.of(ILLEGAL_STATE);
+    }
+
+    protected abstract void handleDone(Port onPort);
+
 
     protected Port createPort(ActorNode<?> remote) {
         earliestReachableNodeBirth = Math.min(this.earliestReachableNodeBirth, remote.birthTime);
